Add files via upload
This project is based on Exercise 3.9 in the textbook UNIX SYSTEMS Programming by Robbins and Robbins.
This program is an expansion of the fan structure as written in Program 3.2 from the textbook through the use of a batch processing facility called 'runsim'. The goal of the assignment as given by Prof. Bhatia is "to become familiar with the environment in Hoare while practicing system calls."